requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
esdk-obs-python==3.20.11
torch==1.10.1
numpy==1.19.3
opencv-python==4.4.0.46
albumentations==1.0.0
torchvision==0.11.2
requests==2.24.0
celery==5.1.2
redis==3.5.3
PyYAML==5.3.1
scikit-learn==0.23.2
scikit-image==0.18.3
pandas==1.1.3
cos-python-sdk-v5==1.9.10
tqdm==4.62.3